Damage System

+++ godot c sharp game/Script/Object/Player/PlayerGrowth.cs
@@ -13,9 +13,9 @@ public class PlayerGrowth {
   }
 
   public void Calculate() {
-    props.Hp1 = (long)(PlayerClass.HpRatio + (props.Level / (1 + PlayerClass.HpBase) + PlayerClass.HpAcc) * props.Level);
-    props.Mp1 = (long)(PlayerClass.MpRate + (props.Level / (1 + PlayerClass.MpBase) + PlayerClass.MpAcc) * PlayerClass.MpRate * props.Level);
-    props.Exp1 = (long)(props.Level * ExpConst * ExpFactor * (props.Level * 1.1));
+    props.Hp1 = (int)(PlayerClass.HpRatio + (props.Level / (1 + PlayerClass.HpBase) + PlayerClass.HpAcc) * props.Level);
+    props.Mp1 = (int)(PlayerClass.MpRate + (props.Level / (1 + PlayerClass.MpBase) + PlayerClass.MpAcc) * PlayerClass.MpRate * props.Level);
+    props.Exp1 = (int)(props.Level * ExpConst * ExpFactor * (props.Level * 1.1));
     props.Ac0 = CalcValue(0.84m, PlayerClass.DefRate, PlayerClass.DefAcc);
     props.Ac1 = CalcValue(0.96m, PlayerClass.DefRate, PlayerClass.DefAcc, true);
     props.Mac0 = CalcValue(0.1m, PlayerClass.DefRate, PlayerClass.DefAcc);
@@ -24,7 +24,7 @@ public class PlayerGrowth {
     props.Sc1 = CalcValue(0m, PlayerClass.AtkRate, PlayerClass.AtkAcc, true);
   }
 
-  private long CalcValue(
+  private int CalcValue(
           decimal c,
           decimal rate,
           decimal acc,
@@ -35,6 +35,6 @@ public class PlayerGrowth {
     var value = Math.Floor(_lv * rate + acc * rate + 1);
     if (boMax) value += _lv * acc;
 
-    return (long)Math.Floor(value);
+    return (int)Math.Floor(value);
   }
 }
